Move legacy methods & SuperblockHeightValidator into separate files
