+++ divi/src/SuperblockHelpers.h
@@ -9,39 +9,11 @@
 class CBlockRewards;
 class CChainParams;
 
-namespace Legacy
-{
-    bool IsValidLotteryBlockHeight(int nBlockHeight,const CChainParams& chainParams);
-    bool IsValidTreasuryBlockHeight(int nBlockHeight,const CChainParams& chainParams);
-    int64_t GetTreasuryReward(const CBlockRewards &rewards, const CChainParams& chainParams);
-    int64_t GetCharityReward(const CBlockRewards &rewards, const CChainParams& chainParams);
-    int64_t GetLotteryReward(const CBlockRewards &rewards, const CChainParams& chainParams);
-    CBlockRewards GetBlockSubsidity(int nHeight, const CChainParams& chainParams);
-    CAmount GetFullBlockValue(int nHeight, const CChainParams& chainParams);
-};
-
 bool IsValidLotteryBlockHeight(int nBlockHeight);
 bool IsValidTreasuryBlockHeight(int nBlockHeight);
 
 CBlockRewards GetBlockSubsidity(int nHeight);
 
-class SuperblockHeightValidator: public I_SuperblockHeightValidator
-{
-private:
-    const CChainParams& chainParameters_;
-    int transitionHeight_;
-    int superblockCycleLength_;
-
-public:
-    SuperblockHeightValidator(const CChainParams& chainParameters);
-    int getTransitionHeight() const;
-    const CChainParams& getChainParameters() const;
-
-    virtual int GetTreasuryBlockPaymentCycle(int nBlockHeight) const;
-    virtual int GetLotteryBlockPaymentCycle(int nBlockHeight) const;
-    virtual bool IsValidLotteryBlockHeight(int nBlockHeight) const;
-    virtual bool IsValidTreasuryBlockHeight(int nBlockHeight) const;
-};
 class BlockSubsidyProvider: public I_BlockSubsidyProvider
 {
 private:
